renames form components for clarity
